Added network layer and partial implementation of database

+++ app/src/main/java/com/spardo/lealTransactions/data/network/dto/BranchDTO.kt
@@ -3,6 +3,6 @@ package com.spardo.lealTransactions.data.network.dto
 import com.google.gson.annotations.SerializedName
 
 data class BranchDTO(
-    @SerializedName("id") val id: Int?,
+    @SerializedName("id") val id: Long?,
     @SerializedName("name") val name: String?
 )
