+++ app/src/main/java/com/spardo/lealTransactions/domain/models/CommerceModel.kt
@@ -1,8 +1,8 @@
 package com.spardo.lealTransactions.domain.models
 
 data class CommerceModel(
-    val id: Int,
+    val id: Long,
     val name: String,
-    val valueToPoints: Int,
+    val valueToPoints: Double,
     val branches: List<BranchModel>
 )
